+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased (master)
 
+### Fixes
+
+* Fix check for passed `pending` tests
+
 ### 0.3.0 (2022-01-15)
 
 ### New Features